test: move http interface test to test_config.py
WAZO-4352OD
