+++ integration_tests/suite/test_config.py
@@ -1,10 +1,11 @@
-# Copyright 2024 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2024-2026 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 import pytest
+import requests
 from hamcrest import assert_that, equal_to, has_entry
 
-from .helpers.base import APIIntegrationTest
+from .helpers.base import VALID_TOKEN, APIIntegrationTest
 
 FAKE_EVENT = {'data': 'Event: foo\r\nAnswerToTheUniverse: 42\r\n\r\n'}
 
@@ -45,3 +46,27 @@ class TestConfigAPI(APIIntegrationTest):
         debug_false_config = self.amid.config()
         assert_that(debug_false_patched_config, equal_to(debug_false_config))
         assert_that(debug_false_config, has_entry('debug', False))
+
+    def test_that_empty_body_returns_400(self) -> None:
+        port = self.asset_cls.service_port(9491, 'amid')
+        url = f'http://127.0.0.1:{port}/1.0/config'
+
+        headers = {
+            'X-Auth-Token': VALID_TOKEN,
+        }
+
+        response = requests.patch(
+            url,
+            headers=headers,
+            data='',
+            verify=False,
+        )
+        assert response.status_code == 400
+
+        response = requests.patch(
+            url,
+            headers=headers,
+            data=None,
+            verify=False,
+        )
+        assert response.status_code == 400
